+++ src/TextCapturer.ts
